hunyuan3d/requirements-minimal.txt
# Minimal Requirements for Hunyuan3D Geometry-Only Mode
# This installs ONLY what's needed for basic 3D mesh generation without textures

# Core dependencies (essential)
numpy>=1.21.0,<2.0  # Must be <2.0 for shapely/trimesh compatibility
pillow>=9.0.0
opencv-python>=4.7.0

# 3D mesh processing (lightweight)
trimesh>=3.15.0
pygltflib>=1.16.0
meshio>=5.3.0

# Web server and API (required for server)
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Utilities
tqdm
pyyaml>=6.0
requests>=2.28.0
psutil>=5.9.0

# Basic image processing (for blueprint analysis)
scikit-image>=0.20.0
imageio>=2.28.0

# Python environment management
python-dotenv>=1.0.0

# Note: This excludes:
# - torch/torchvision (heavy AI models)
# - transformers/diffusers (AI inference)
# - xformers (memory optimization)
# - rembg/onnxruntime (background removal AI)
# - open3d/pymeshlab (advanced 3D processing)
# - gradio (web UI)
# These can be added later when needed for full AI functionality
